Fix Travis CI integration
* General:
  - Fix integration with Travis CI. Handle Travis CI webhook correctly.
* Models:
  - Delete many to many relationships between `Notification` and `Content`.
  - Delete many to many relationships between `Notification` and
    `ImportTask`.
  - Drop `Notification.messages` fields as not used.
  - New relations `Notification.last_import`, `Notification.repository`.
* Views:
  - Extract notification views to standalone module
    `galaxy/serializers/notification.py`.
* Serializers:
  - Extract notification serializers to standalone module
    `galaxy/views/notification.py`
