+++ galaxy/api/serializers/__init__.py
@@ -19,6 +19,7 @@ from .content import *             # noqa
 from .content_block import *       # noqa
 from .content_type import *        # noqa
 from .namespace import *           # noqa
+from .notification import *        # noqa
 from .provider import *            # noqa
 from .provider_source import *     # noqa
 from .repository_source import *   # noqa
