+++ galaxy/api/serializers/serializers.py
@@ -33,8 +33,6 @@ from galaxy.main.models import (Platform,
                                 Content,
                                 ImportTask,
                                 RepositoryVersion,
-                                NotificationSecret,
-                                Notification,
                                 Subscription,
                                 Stargazer
                                 )
@@ -55,8 +53,6 @@ __all__ = [
     'PlatformSearchSerializer',
     'RepositoryVersionSerializer',
     'TopContributorsSerializer',
-    'NotificationSecretSerializer',
-    'NotificationSerializer',
     'ImportTaskSerializer',
     'ImportTaskLatestSerializer',
     'RoleTopSerializer',
@@ -460,97 +456,6 @@ class TopContributorsSerializer(serializers.BaseSerializer):
             'namespace': obj['namespace'],
             'role_count': obj['count']
         }
-
-
-class NotificationSecretSerializer(BaseSerializer):
-    secret = serializers.SerializerMethodField()
-
-    class Meta:
-        model = NotificationSecret
-        fields = (
-            'id',
-            'owner',
-            'github_user',
-            'github_repo',
-            'source',
-            'secret',
-            'created',
-            'modified',
-            'active'
-        )
-
-    def get_url(self, obj):
-        if obj is None:
-            return ''
-        elif isinstance(obj, NotificationSecret):
-            return reverse('api:notification_secret_detail', args=(obj.pk,))
-        else:
-            return obj.get_absolute_url()
-
-    def get_secret(self, obj):
-        # show only last 4 digits of secret
-        last = ''
-        try:
-            last = obj.secret[-4:]
-        except Exception:
-            pass
-        return '******' + last
-
-
-class NotificationSerializer(BaseSerializer):
-    class Meta:
-        model = Notification
-        fields = (
-            'id',
-            'owner',
-            'source',
-            'github_branch',
-            'travis_build_url',
-            'travis_status',
-            'commit_message',
-            'committed_at',
-            'commit',
-            'messages',
-            'created',
-            'modified'
-        )
-
-    def get_url(self, obj):
-        if obj is None:
-            return ''
-        elif isinstance(obj, Notification):
-            return reverse('api:notification_detail', args=(obj.pk,))
-        else:
-            return obj.get_absolute_url()
-
-    def get_summary_fields(self, obj):
-        if obj is None:
-            return {}
-        d = super(NotificationSerializer, self).get_summary_fields(obj)
-        d['owner'] = OrderedDict([
-            ('id', obj.owner.id),
-            ('username', obj.owner.username)
-        ])
-        d['roles'] = [OrderedDict([
-            ('id', r.id),
-            ('namespace', r.namespace),
-            ('name', r.name)
-        ]) for r in obj.roles.all()]
-        d['imports'] = [OrderedDict([
-            ('id', t.id)
-        ]) for t in obj.imports.all()]
-        return d
-
-    def get_related(self, obj):
-        if obj is None:
-            return {}
-        res = super(NotificationSerializer, self).get_related(obj)
-        res.update(dict(
-            roles=reverse('api:notification_roles_list', args=(obj.pk,)),
-            imports=reverse('api:notification_imports_list', args=(obj.pk,)),
-            owner=reverse('api:user_detail', args=(obj.owner.id,)),
-        ))
-        return res
 
 
 class ImportTaskSerializer(BaseSerializer):
