+++ galaxy/api/views/__init__.py
@@ -20,6 +20,7 @@ from .content import *              # noqa
 from .content_block import *        # noqa
 from .content_type import *         # noqa
 from .namespace import *            # noqa
+from .notification import *         # noqa
 from .repository import *           # noqa
 from .provider_source import *      # noqa
 from .repository_source import *    # noqa
